Use Go 1.24 os.Root instead of filepath-securejoin
All the features of securejoin used in k/k can be replaced by
equivalents in Go 1.24's new os.Root() feature.
